Fix 32-bit vs 64-bit integer overflows and GUI thread safety
- Upgrade sample counts and file size tracking in input.h/c, mp4write.h/c, and encode_engine.h/c to 64-bit types.
- Fix g_mp4.buffersize typing and memory allocation growth upper bounds in mp4write.c.
- Safely clamp 64-bit sample counts for standard 32-bit MP4 container atoms.
- Fix Win32 GUI thread safety race condition in maingui.c by snapshotting parameters for the background EncodeFile thread.
- Improve progress bar math and edit control input sanitization in maingui.c.

+++ frontend/encode_engine.h
@@ -127,7 +127,7 @@ typedef struct {
 
 typedef struct {
     uint32_t frame_count;
-    uint32_t sample_count;
+    uint64_t sample_count;
     uint32_t max_bitrate;
     uint32_t avg_bitrate;
     uint32_t max_frame_size;
